Log a warning rather than throwing an error when groupdata not found
The snapshot package throws a pretty uninformative "GroupId not
available" error when something goes wrong loading data. This updates
it to instead log a warning that includes additional data about the
request, so that we can better investigate what's really going on,
and ops won't have to content with as much logspam in the meantime.

Bug: T158657
Change-Id: Ie05f691e97b00e4ba55b9d22fee1ac97025142e1