pipeline: Fix execution configuration
The `execution` field should be given as a list of lists (each being
stage names that define an execution graph branch), not a one
dimensional list as was given. Additionally, the attempted execution
graph that was given is the default (one branch comprised of all defined
stages in a serial order) and can be omitted.

Bug: T233316
Change-Id: Ia599ccafbd7700ecc62bbf97c42e0427fc57b3dc