Migrate mocha tests to jest
Did this so that I can use the jest mock library in the snapshot
test.

Patch was created using the automated transformation `npx
jest-codemods`, and inlining the test timeouts.

TODO: The server can't be run concurrently so jest is limited to 1
worker.  Maybe we can get around that limitation by running the
servers on multiple ports, or by rewriting as more unit-y tests.

Change-Id: I467a8c4a2875917149a771d6c5abb8926a2ae84e