Accept full URL in domain parameter
This makes it possible to connect back to a MediaWiki on a
non-standard port, for example when MediaWiki is hosted in a sibling
docker-compose cluster.

Still makes an assumption about api.php, so the URL support
is incomplete and ignores any path or fragment.

Since the port is stripped before checking domains, we can leave port
out of the example allowDomains.

Bug: T301769
Change-Id: I5ba7f49902b2404ba219b318f4db1a589438513b