Update module lock files to current versions
This hopefully helps to keep diffs of upcomming changes small that
change dependencies in the modules and spot unrelated or missing
changes more easily.

With this you should be able to run the clean modules script for a
clean re-install of the dependencies and don't end up with uncommited
changes due to dirty lock file diffs. ( at least for now )

Change-Id: I4c185f7c7a665ab8d5d379fe4e90cc67ecc4d85d