Bump eslint-config-wikimedia to 0.20.0
This is the latest major version before node12.

Eslint is now also applying for json files so a big part of the
changes are whitespace there. Another big chunk is about jsdoc that
also includes whitespace or autogenerated parameter documentation.

Mainly autofixes, only in some cases the jsdoc autogeneration did
strange things. Fixed that.

Apart from that we have:
- adding 'use strict' to all js under the server rule set
- tweaking the exceptions for jsdoc ( although just warnings )
- minor fix to one return type jsdoc
- minor max len fix

Change-Id: I138ab879c0b47efb2f29f794a345296f0c601836