Fix a series of type-reated ESLint errors and warnings
In a lot of cases I go for the most generic type, e.g. {Object} or
even {*} when I can't be 100% sure what the exact type might be.
This can be improved in later patches. If you can't verify a type and
that blocks this patch, let's make it more generic.

Change-Id: Idd2446009627ddaabc3c0c1c7474ad9a896a64bb