Log errors at higher severity
Don't rely on metrics in place of errors—production is configured to
send metrics to statsd, which is nearly one-dimensional and strips
the useful metadata.  Either way, we want logstash to include all the
errors so we can do consistent measurement there.

Any error produced using this pattern was not landing in logstash
because of the issue:

    throw new Err( 'foo' ).metrics( 'err.req.foo' );

Bug: T322616
Change-Id: Ica9bd15650d6d06548168614cc662a4e53f6f750