Remove unused tilelive modules
There's no reference for `autogen` and `substantial` in the code. Also
from smoke testing it seems that the code is not reached from any input
I'm aware of.

See https://en.wikipedia.beta.wmflabs.org/wiki/KartotherianSmoketests
for a list of my tests.

`tilelive-switch` introduced `mirror` that's not even loaded in the
config files anymore and also not referenced in any other place. It
was used when migrating to Tegola. But that's done now.

Also cleaning up the readme and removing the reference for the also
not used demultiplexer.

Change-Id: I8441d5655ce4dd6c62ed4697850945bcc5262d0e