Move flattening for overlay scale further down
Abaculus is fine with an unset scale. It internally defaults to 1.

Generally valid values are defined in the source configuration and
checked with `core.validateScale`.

Overlays further reduce the number of valid scale values. The reason
for that seem to be the maki icons because they only support scaling
1 or 2.

The pure stiched base map could be scaled differently. The change
reflects that fact better.

Further down the metrics assume a set scale so that was fixed.

Change-Id: Ic1e2dfd84de457a48abfb18c596b46525bc3e6a3