Streamline scale route parsing
Retrieving the scale is reduced to getting the number between
the `@` and `x`. Now this does not have to be cleaned up later.

The optional scale is reflected in tiles.js with an question mark.

The format is never optional and can not be empty. There's no reason
why the route should listen to empty formats.

Note, that this now also better reflects what is written in the
readme.

Change-Id: I7c64b72f84ce1fc18fd923bbb7bc7b951832c38e