Uncomment tile handler spec
This will also trigger the corresponding spec test that pokes the
endpoint and checks if the result is not failing.

The two specs still commented out, that would test the scale end-
points on tile and snapshot, are currently failing the test.

Try to deal with that in a follow up.

Change-Id: I89cf15d9a370ca5475a002265093397720cc4cc3