Adjust tile data integration test to new values
Data here is coming from the https://maps.wikimedia.org/osm-pbf/
endpoint. So all test work with the same (live) data as source.

Seems that that map got an update and therefore the id's changed.
On the long run it might not make sense to keep that test. Or we
should make it more simple.

Change-Id: I87ffe38aba76f34a98886d1af87a45a59cceffbd