Re-enable loader tests
Seems these tests where not executed for some time. Maybe related
to moving to jest. Now they will run again.

Two failing tests are commented out. I'm not sure if it's worth it
to fix the XmlLoader test. AFAIK this is a dead module. The source
configuarion where these module loaders are used is in Yaml and
from poking the XmlLoader is never exectued anywhere.

Change-Id: Ie0022eb5bd6f7e0cd2cf3e345881232cbd9bdd21