LanguagePicker.js: Remove fallback Rule 3
Remove language fallback Rule 3, as it will cause a random language in the same script to be displayed even if the default name would be more appropriate (for example, for locations in the US where the native name is English, Serbian names are displayed instead of the native English names because the Serbian entries have the "-Latn" suffix). Displaying the default native name is likely preferable to the name in a random language with the same character set. Also removed associated test case from langPicker.test.js.

Bug: T195318
Change-Id: I23d4e1f41bfd1e8ca129b2438a00ffafdcaf5bdb