Remove mapnik build step in favor of a base image
Remove the mapnik build step in favor of the same done in a base
(build) image, that is re-used every time a change is filed
for Kartotherian (more efficient for devs, it cuts the wait time
from minutes to less than one).

The final test/production images are also changed, to use runtime
dependency libs rather than -dev ones (better in terms of security
exposure and space used on disk).

Bug: T327396
Depends-On: https://gerrit.wikimedia.org/r/c/operations/docker-images/production-images/+/1100070
Change-Id: I9c5cab492123e84b59f81d4c304ea11a4de95f78