blubber: add config for jemalloc
The Kartotherian container running on K8s seems to show a slow but
steady increase of RSS memory used (see T386926#10624612). The heap
size doesn't grow, so it may be a memory leak or Nodejs not being
able to reclaim "external" allocated memory (for example, from
the mapnik C++ library). A lot of people suggest to try jemalloc first,
especially in these use cases, because it could really improve
the allocation/deallocation behavior and performances (avoiding
as much as possible fragmentation, that sometimes may appear as leak).
It is a change that is not guided by solid data,
but it would be nice anyway to check if changing malloc() implementation
we could improve the current situation.

Bug: T386926
Change-Id: I10eed630d8fab6234a9c40a69d3e10c4da27d186